533 Move BBL optionsets into their own file
