+++ client/app/optionsets/bbl.js
@@ -1,0 +1,26 @@
+export const BOROUGHS_OPTIONSET = {
+  BRONX: {
+    code: 717170000,
+    label: 'Bronx',
+  },
+  BROOKLYN: {
+    code: 717170002,
+    label: 'Brooklyn',
+  },
+  MANHATTAN: {
+    code: 717170001,
+    label: 'Manhattan',
+  },
+  QUEENS: {
+    code: 717170003,
+    label: 'Queens',
+  },
+  STATEN_ISLAND: {
+    code: 717170004,
+    label: 'Staten Island',
+  },
+  CITYWIDE: {
+    code: 717170005,
+    label: 'Citywide',
+  },
+};
